# Conflicts:
#	src/main/java/com/kodigo/bank/entity/BankLoan.java

+++ src/main/java/com/kodigo/bank/entity/LoanPayment.java
@@ -12,7 +12,7 @@ import java.math.BigDecimal;
 public class LoanPayment {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int paymentId;
+    private long paymentId;
     private BigDecimal amount;
     private LocalDate date;
